Fixes
- Fixed max party size permissions.
(https://github.com/rlf/uSkyBlock/issues/430)
- Fixed random diagonal lines being generated through islands.
- Re-added the chunk generator memory improvement fix.
- Fixed offline player islands commands not working.
