--- src/main/java/us/talabrek/ultimateskyblock/command/admin/AbstractPlayerInfoCommand.java
@@ -1,34 +1,0 @@
-package us.talabrek.ultimateskyblock.command.admin;
-
-import org.bukkit.command.CommandSender;
-import us.talabrek.ultimateskyblock.command.common.AbstractUSBCommand;
-import us.talabrek.ultimateskyblock.player.PlayerInfo;
-import us.talabrek.ultimateskyblock.uSkyBlock;
-
-import java.util.Map;
-
-import static us.talabrek.ultimateskyblock.util.I18nUtil.tr;
-
-/**
- * Command that has <code>player</code> as first argument, and uses playerInfo.
- */
-public abstract class AbstractPlayerInfoCommand extends AbstractUSBCommand {
-    protected AbstractPlayerInfoCommand(String name, String permission, String description) {
-        super(name, permission, "player", description);
-    }
-    protected abstract void doExecute(CommandSender sender, PlayerInfo playerInfo);
-    @Override
-    public boolean execute(CommandSender sender, String alias, Map<String, Object> data, String... args) {
-        if (args.length > 0) {
-            String playerName = args[0];
-            PlayerInfo playerInfo = uSkyBlock.getInstance().getPlayerInfo(playerName);
-            if (playerInfo != null) {
-                data.put("playerInfo", playerInfo);
-                doExecute(sender, playerInfo);
-                return true;
-            }
-            sender.sendMessage(tr("\u00a7eInvalid player {0} supplied.", args[0]));
-        }
-        return false;
-    }
-}
