+++ src/main/java/us/talabrek/ultimateskyblock/command/admin/GotoIslandCommand.java
@@ -1,18 +1,16 @@
 package us.talabrek.ultimateskyblock.command.admin;
 
-import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import us.talabrek.ultimateskyblock.player.PlayerInfo;
 import us.talabrek.ultimateskyblock.uSkyBlock;
-
 import static us.talabrek.ultimateskyblock.util.I18nUtil.tr;
 
 /**
  * Teleports to the player's island.
  */
-public class GotoIslandCommand extends AbstractPlayerInfoCommand {
+public class GotoIslandCommand extends AbstractAsyncPlayerInfoCommand {
     private final uSkyBlock plugin;
 
     public GotoIslandCommand(uSkyBlock plugin) {
@@ -21,7 +19,7 @@ public class GotoIslandCommand extends AbstractPlayerInfoCommand {
     }
 
     @Override
-    protected void doExecute(CommandSender sender, PlayerInfo playerInfo) {
+    protected void doExecute(final CommandSender sender, final PlayerInfo playerInfo) {
         if (!(sender instanceof Player)) {
             sender.sendMessage(tr("\u00a74Only supported for players"));
         }
